Move some model utils into vision file
